requirements.txt
Appium-Python-Client>=2.10.2
selenium>=4.17.2